vtkXML*: Use 64-bit integer types where appropriate
Use size_t for memory buffer sizes and vtkTypeInt64 for file positions
and offsets exposed in header files.  Use vtkTypeUInt64 for sizing and
indexing of whole data.  The latter will be necessary to read small
portions of data requiring 64-bit indexes on 32-bit architectures.
Use std::streampos for stream seek operations but avoid exposing it
in header files because its type can vary on some platforms depending
on preprocessor definitions for large file support.

+++ IO/XML/vtkXMLRectilinearGridWriter.cxx
@@ -99,7 +99,7 @@ vtkXMLRectilinearGridWriter::CreateExactCoordinates(vtkDataArray* a, int xyz)
     {
     // Create a subset of the coordinates array.
     int components = a->GetNumberOfComponents();
-    int tupleSize = components*this->GetWordTypeSize(a->GetDataType());
+    size_t tupleSize = components*this->GetWordTypeSize(a->GetDataType());
     vtkDataArray* b = a->NewInstance();
     b->SetNumberOfComponents(components);
     b->SetName(a->GetName());
